Small changes in a couple of methods and tests which were failing in Jenkins.

+++ war/src/main/java/nl/mpi/lamus/configuration/Configuration.java
@@ -44,5 +44,5 @@ public interface Configuration {
     
     public File getRelaxedTypeCheckConfigFile();
     
-    public int getTypeReCheckSizeLimit();
+    public long getTypeReCheckSizeLimit();
 }
